Impl. RichText block
