+++ src/bundle/Resources/public/css/alloyeditor/alloyeditor-ez.css
@@ -353,7 +353,6 @@
 }
 
 .ae-ui .ae-toolbar-add {
-  border: solid 2px transparent;
   padding: 0;
 }
