+++ src/bundle/Resources/public/scss/fieldType/edit/_ezrichtext.scss
@@ -3,96 +3,5 @@
         display: none;
     }
 
-    .ez-data-source__richtext {
-        background-color: $ez-white;
-        min-height: 100px;
-        border: 1px solid rgba(0, 0, 0, 0.15);
-        border-radius: 0.25rem;
-        display: inline-block;
-        width: 100%;
-        padding: 0.5rem;
-
-        .is-block-focused,
-        .cke_widget_wrapper.cke_widget_focused > .cke_widget_element {
-            outline: 2px dashed $ez-ground-base-dark;
-            outline-offset: 1px;
-        }
-
-        &.is-invalid {
-            background: $ez-color-warning-pale;
-            border-color: $ez-color-danger;
-        }
-
-        blockquote {
-            margin: 1rem 2.5rem;
-        }
-    }
-
-    [data-ezalign='right'] {
-        float: right;
-    }
-
-    [data-ezalign='left'] {
-        float: left;
-    }
-
-    [data-ezalign='center'] {
-        text-align: center;
-    }
-
-    [data-ezelement='ezconfig'] {
-        display: none;
-    }
-
-    [data-ezelement='ezembed'] {
-        background: $ez-ground-base;
-        border: 1px solid $ez-ground-base-dark;
-        font-size: 1.1rem;
-        display: inline-block;
-        margin: 0.5rem 0.3rem;
-        padding: 0.3rem;
-        box-sizing: border-box;
-        line-height: 1;
-        min-height: auto;
-
-        .ez-embed-content {
-            margin-bottom: 0;
-        }
-    }
-
-    pre {
-        padding: calculateRem(10px);
-        min-height: calculateRem(43px);
-        background-color: $ez-ground-base-pale;
-        border: calculateRem(1px) solid $ez-ground-base-medium;
-        word-wrap: initial;
-    }
-
-    p {
-        min-height: calculateRem(24px);
-    }
-
-    h1 {
-        min-height: calculateRem(33px);
-    }
-
-    h2 {
-        min-height: calculateRem(28px);
-    }
-
-    h3 {
-        min-height: calculateRem(26px);
-    }
-
-    h4 {
-        min-height: calculateRem(24px);
-    }
-
-    h5 {
-        min-height: calculateRem(21px);
-    }
-
-    h6 {
-        min-height: calculateRem(20px);
-    }
+    @include richtext-field();
 }
