+++ src/bundle/Resources/views/content/content_edit/parts/javascripts.html.twig
@@ -6,6 +6,8 @@
     '@EzPlatformAdminUiBundle/Resources/public/js/scripts/fieldType/base/base-field.js'
     '@EzPlatformAdminUiBundle/Resources/public/js/scripts/fieldType/base/base-file-field.js'
     '@EzPlatformAdminUiBundle/Resources/public/js/scripts/fieldType/base/base-preview-field.js'
+    '@EzPlatformAdminUiBundle/Resources/public/js/scripts/fieldType/base/base-rich-text.js'
+    '@EzPlatformAdminUiBundle/Resources/public/js/scripts/fieldType/alloyeditor/custom.tags.js'
     '@EzPlatformAdminUiBundle/Resources/public/js/scripts/fieldType/*'
     '@EzPlatformAdminUiBundle/Resources/public/js/scripts/sidebar/extra.actions.js'
 %}
